Added bad code to render x2 and all for targeted cards
RenderEntitiesSystem now has some hackish code to render "x2" and "all" hands. This needs to be completely changes, likely working with a new version of multipositions where we can spin up some new entities. Maybe multiposition entity is bad and what we really need are some sprite clone entities that we add and remove or something, not sure. Like, an entity with a clone component or something.