Started to consolidate wrapper logic into systems
Started with cursor methods mainly to try to remove ones we don't need as well as find ones that are only used in a particular system or a couple systems and move the logic into the systems themselves. This may cause some redundancy but at the same time, it gets us closer to the ideal design for Ashley.