Refactoring overly complicated card and turn action fields
Trying to simplify and generalize how cards and turn actions are processed by removing a lot of sweeping assumptions. Added a generic "ownerID" component that says that an entity belongs to something. This can be processed based on the context of what that entity is. Will want to add a similar type of component that points to the current "sprite entity" for a given component, so that a player can point to an avatar without us knowing we're dealing with a player or not.