Added DiscardZoneSystem
Will likely rename and maybe redo how we handle this system, but the idea is just by adding a "toZone" comp to an entity we should be able to handle what happens when it's moved to that zone. This is somewhat a replacement or at least an interum system for "ZoneChangeSystem" since a zone change is still too much logic for the average entity system to deal with, what with getting zones by id and setting and all that junk. Also, TurnAction and CardComp are becoming increasingly coupled and it might just be better to combine them into the same component.