Starting to use new target system and new custom systems
Added custom systems that can store and query more entity lists than just an iterating system. This will eventually (hopefully) replace all global ids stored in favor or "first of this type" queries attached to specific systems.

Added new cursor targetting system that will make anything with a "targetable" comp be targetable by the cursor. It still has more code than I'd like but the reason to update the system is to try to make it more general and not reason on specific zones so much, only zones relative to the cursor.