Refactored wrappers
Renamed CompWrapper classes to CompUtil, so now you do Comp.util(v).damage(100) and the like to get access to those extra methods. CompUtil is meant to hold bread and butter book keeping and number crunching code, not shortcuts to other components or complicated linking logic. That should be help be the system's families and collections. If it turns out util methods are and only ever will be used in one system, they should be moved directly into that system and taken away from utils.