Added CardOwnerComp and Removed Enemy Phases
Added CardOwnerComp to denote which entities have cards associated with them. This can be used in conjunction with the new CardOwnerActionSystem to do things like fill hands, draw cards, reset cards at the beginning of battle ect. Now enemies are handled just like players, with cards processed in real time. Added component that gives an enemy a wait time. Could eventually be seeded by the enemies speed stat at the start of battle or something, but for now it stands alone.