Added event comps
Refactored a bit to add "Event" comps, ones that make it exactly one tick around and then are cleared off all components. Handy for sending events like zone changes and whatnot. There are probably tons of interactions that could be refactored to use this, but right now it's used to communicate cursor state and respond to entities being selected and deselected by the cursor.