Fixed pending bug and added confuse and skip
Added new status effects for confuse and skip as well as fixed some complicated logic where we thought we were getting pending values but were actually getting "actual" values. Not completely happy with how the skip and confuse are implemented, with confuse in particular being a problem for selecting targets. Maybe it should be applied when a turn is actually used instead of during targetting.